+++ app_scheduler/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class App_SchedulerConfig(AppConfig):
+    name = 'app_scheduler'
