+++ app_scheduler/serializers.py
@@ -1,0 +1,14 @@
+from django.contrib.auth.models import User, Group
+from rest_framework import serializers
+
+
+class UserSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = User
+        fields = ['organization', 'display_name', 'student_id', 'user_schedule']
+
+
+class GroupSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Group
+        fields = ['team_id', 'team_schedule']
