+++ app_scheduler/views.py
@@ -1,10 +1,31 @@
 from django.shortcuts import render
 
-# Create your views here.
-from django.http import HttpResponse
-from .scraping import getPageSource, getTime, exportAsDict, makeList, changeData, beRowData
-from scheduler.models import  User, Org
+from django.http import HttpResponse, JsonResponse
+from django.views.decorators.csrf import csrf_exempt
+from rest_framework import viewsets
+from rest_framework.parsers import JSONParser
 
+from .scraping import getPageSource, getTime, exportAsDict, makeList, changeData, beRowData
+from django.contrib.auth.models import User, Group
+from app_scheduler.serializers import UserSerializer, GroupSerializer
+from .models import Organization
+
+# Create your views here.
+
+class UserViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows users to be viewed or edited.
+    """
+    queryset = User.objects.all().order_by('-date_joined')
+    serializer_class = UserSerializer
+
+
+class GroupViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows groups to be viewed or edited.
+    """
+    queryset = Group.objects.all()
+    serializer_class = GroupSerializer
 
 def index(request):
     #indexを読み込んでデフォルトの画面表示
@@ -31,11 +52,13 @@ def execute(request):
         aft_class = beRowData(aft_class)
 
         # 書き出しサイズがオーバーフローを起こすため10進数に変換して書き込む
-        db_org = Org(
+        """
+        db_org = Organization(
             share_pre = int(pre_class, 2),
             share_aft = int(aft_class, 2),
         )
         database_p.save()
+        """
 
     return render(request, "scheduler/index.html")
 
@@ -43,7 +66,7 @@ def execute(request):
 def makeorg(request):
     if request.method == 'POST':
         team_id = request.POST["team_id"] # 任意のteam_id
-        db_org = Org(team_id=team_id)
+        db_org = Organization(team_id=team_id)
         db_org.save()
 
     return render(request, "scheduler/makeorg.html")
@@ -51,9 +74,9 @@ def makeorg(request):
 
 def joinorg(request):
     if request.method == 'POST':
-        if request.filter(Org.objects.get("team_id").exists()):
+        if request.filter(Organization.objects.get("team_id").exists()):
             print("ok")
-        #team_id = request.POST["org"] # 参加するチームのteam_id
+        team_id = request.POST["org"] # 参加するチームのteam_id
         display_name = request.POST["display_name"] #チーム内での任意の表示名
 
         db_usr = User(team_id=team_id, display_name=display_name)
