+++ scheduler/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for univScraping project.
+Django settings for scheduler project.
 
 Generated by 'django-admin startproject' using Django 2.2.
 
@@ -37,9 +37,21 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'rest_framework',
     "bootstrap4",
-    "scheduler",
+    "app_scheduler",
 ]
+
+REST_FRAMEWORK = {
+    # Use Django's standard `django.contrib.auth` permissions,
+    # or allow read-only access for unauthenticated users.
+    'DEFAULT_PERMISSION_CLASSES': [
+        'rest_framework.permissions.AllowAny',
+    ],
+    # 'EXCEPTION_HANDLER': 'scheduler.app_scheduler.utils.custom_exception_handler',
+    'DEFAULT_PAGENATION_CLASS': 'rest_framework.pagination.PageNumberPagination',
+    'PAGE_SIZE': 10
+}
 
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
@@ -51,7 +63,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'univScraping.urls'
+ROOT_URLCONF = 'scheduler.urls'
 
 TEMPLATES = [
     {
@@ -69,7 +81,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'univScraping.wsgi.application'
+WSGI_APPLICATION = 'scheduler.wsgi.application'
 
 
 # Database
