+++ scheduler/urls.py
@@ -1,12 +1,31 @@
-from django.urls import path
+"""scheduler URL Configuration
 
-from . import views
+The `urlpatterns` list routes URLs to views. For more information please see:
+    https://docs.djangoproject.com/en/2.2/topics/http/urls/
+Examples:
+Function views
+    1. Add an import:  from my_app import views
+    2. Add a URL to urlpatterns:  path('', views.home, name='home')
+Class-based views
+    1. Add an import:  from other_app.views import Home
+    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
+Including another URLconf
+    1. Import the include() function: from django.urls import include, path
+    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
+"""
+from django.contrib import admin
+from django.urls import include, path
+from rest_framework import routers
+from app_scheduler import views
 
-app_name = "scheduler"
+router = routers.DefaultRouter()
+router.register(r'users', views.UserViewSet)
+router.register(r'groups', views.GroupViewSet)
+router.register(r'organizations', views.GroupViewSet)
 
 urlpatterns = [
-    path('', views.index, name='index'),
-    path('scheduler', views.execute, name="execute"),
-    path("makeorg", views.makeorg, name="makeorg"),
-    path("join", views.joinorg, name="joinorg")
+    path('scheduler/', include("app_scheduler.urls")),
+    path('', include(router.urls)),
+    path('admin/', admin.site.urls),
+    path('api/', include('rest_framework.urls', namespace='rest_framework'))
 ]
