+++ scheduler/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for univScraping project.
+WSGI config for scheduler project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'univScraping.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'scheduler.settings')
 
 application = get_wsgi_application()
